% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData4WR.R
\name{CheckData4WR}
\alias{CheckData4WR}
\title{Check data format for Wasserstein regression}
\usage{
CheckData4WR(X = NULL, Y, qSup)
}
\arguments{
\item{X, Y}{The input data of predictors and responses for \code{\link{WR}}, if both are specified;
Otherwise, if \code{X} is not specified and hence is \code{NULL}, \code{Y} is the input \code{Y} of \code{\link{WAR}}.}

\item{qSup}{Support grid of quantile functions for distributional variables; see \code{\link{WR}} for details.}
}
\value{
A list of the follow elements if \code{X} is specified:
\item{scalarRps}{Logical; whether the responses in \code{Y} are scalars (\code{TRUE}) or distributions (\code{FALSE}).}
\item{anyScalarPdt}{Logical; \code{TRUE} if there exists scalar predictors and \code{FALSE} otherwise.}
\item{anyDistnlPdt}{Logical; \code{TRUE} if there exists distributional predictors and \code{FALSE} otherwise.}
\item{indScalarPdt}{A vector holding the indices of elements in \code{X} corresponding to scalar predictors.}
\item{indDistnlPdt}{A vector holding the indices of elements in \code{X} corresponding to distributional predictors.}
\item{n}{Number of realizations.}
\item{q}{Number of grid points in \code{qSup}.}
}
\description{
Check if there are problems with the form and structure of input data for \code{\link{WR}}.
}
