% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckXpred4WR.R
\name{CheckXpred4WR}
\alias{CheckXpred4WR}
\title{Check prediction set format for Wasserstein regression}
\usage{
CheckXpred4WR(Xpred, resCheckData)
}
\arguments{
\item{Xpred}{A list holding the values of distributional and/or scalar predictors for a held-out set awaiting prediction.}

\item{resCheckData}{A list holding the output of \code{\link{CheckData4WR}}.}
}
\value{
The number of predictions to be made.
}
\description{
Check if there are problems with the form and structure of predictors in the prediction set.
}
