% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIntgWts.R
\name{GetIntgWts}
\alias{GetIntgWts}
\title{Get the weights used for numerical integration given a grid}
\usage{
GetIntgWts(x)
}
\arguments{
\item{x}{A vector holding the grid points where numerical integration is to be taken, the elements of which are in an increasing order.}
}
\value{
A vector holding the weights.
}
\description{
Get the weights used for numerical integration given a grid
}
