% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProj.R
\name{GetProj}
\alias{GetProj}
\title{Project to the log image space}
\usage{
GetProj(LogYhat, LogSup, optns = list())
}
\arguments{
\item{LogYhat}{A \eqn{p}-by-\eqn{n} matrix holding \eqn{n} fitted/predicted log mapped distributions 
evaluated on a common grid given by \code{LogSup} of length \eqn{p}.}

\item{LogSup}{A vector of length \eqn{p} holding the common support grid.}

\item{optns}{A list of control parameters specified by \code{list(name=value)}. See \code{\link{WR}} for details.}
}
\value{
A list of the following elements:
\item{Yhat}{A \eqn{p}-by\eqn{n} matrix holding the quantile functions of fitted/predicted distributions.}
\item{outOfLogSpace}{A logical vector holding whether fitted log mapped distributions 
lie out of the log image space (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Obtain projections of fitted/predicted log mapped distributions to the log image space if they are not in the space.
}
