% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hfaction_cpx9}
\alias{hfaction_cpx9}
\title{A subset of the HF-ACTION study data on high-risk non-ischemic heart failure patients}
\format{
A data frame with 1,448 rows and 5 variables:
\describe{
  \item{patid}{patient ID}
  \item{time}{event times (months)}
  \item{status}{event status; 0:censoring, 1:death, 2:hospitalization}
  \item{trt_ab}{treatment indicator: 1: exercise training, 0: usual care}
  \item{age60}{1: 60 years or older, 0: less than 60 years old}
  }
}
\usage{
hfaction_cpx9
}
\description{
These are data on a subgroup of 426 high-risk non-ischemic patients in the HF-ACTION study.
}
\references{
O'Connor, C. M., Whellan, D. J., Lee, K. L., Keteyian, S. J., Cooper, L. S., Ellis, S. J.,
Leifer, E. S., Kraus, W. E., Kitzman, D. W., Blumenthal, J. A. et al. (2009). Efficacy and
safety of exercise training in patients with chronic heart failure: HF-ACTION randomized
controlled trial. Journal of the American Medical Association, 301, 1439--1450.
}
\keyword{datasets}
