% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2boundary.R
\name{map2boundary}
\alias{map2boundary}
\title{Boundary projection for the prediction of log mapped distributions}
\usage{
map2boundary(Logfit, sup)
}
\arguments{
\item{Logfit}{A vector holding a log map to be projected.}

\item{sup}{A vector holding the support grid of \code{Logfit}.}
}
\value{
The projection parameter.
}
\description{
Boundary projection for the prediction of log mapped distributions
}
\references{
\cite{Chen, Y., Lin, Z., & Müller, H.-G. (2021). "Wasserstein regression." Journal of the American Statistical Association, in press.}
}
