% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2qt.R
\name{map2qt}
\alias{map2qt}
\title{Projecting functions to the space of non-decreasing functions.}
\usage{
map2qt(y, lower = NULL, upper = NULL)
}
\arguments{
\item{y}{A function to be projected.}

\item{lower, upper}{Lower and upper bounds of the image space of the function. 
Defaults are both \code{NULL}, corresponding to an unbounded image space, i.e., the entire real line.}
}
\value{
The projected function.
}
\description{
Projecting functions to the space of non-decreasing functions.
}
\references{
\cite{Petersen, A., & Müller, H.-G. (2019). "Fréchet regression for random objects with Euclidean predictors." The Annals of Statistics, 47(2), 691--719.}
}
