% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.WAR.R
\name{predict.WAR}
\alias{predict.WAR}
\title{Predict distributions given a WAR object}
\usage{
\method{predict}{WAR}(object, n.ahead = 1, Ynew = NULL, ...)
}
\arguments{
\item{object}{A WAR object, i.e., an output of \code{\link{WAR}}.}

\item{n.ahead}{Number of steps ahead at which to predict. Default: 1.}

\item{Ynew}{A \eqn{q}-by-\eqn{m} matrix holding the quantile functions of 
\eqn{m} distributions evaluated on \code{object$qSup} to which to apply the prediction, 
where \eqn{m} must be greater than or equal to \code{object$order}.
Default: \code{object$Y}.}

\item{...}{Not used.}
}
\value{
A list of the following elements:
\item{Ypred}{A \eqn{q}-by-\eqn{l} matrix holding the predicted quantile functions of 
\eqn{l} distributions ahead of \code{Ynew}, where \eqn{l} is given by \code{n.ahead}.}
\item{qSup}{The input \code{qSup}}
\item{outOfLogSpace}{A logical vector holding whether initial predicted log mapped distributions 
lie out of the log image space (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Predict distributions given a WAR object
}
