% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.WR.R
\name{predict.WR}
\alias{predict.WR}
\title{Predict responses for a new sample given a WR object}
\usage{
\method{predict}{WR}(object, Xpred, ...)
}
\arguments{
\item{object}{A WR object, i.e., an output of \code{\link{WR}}.}

\item{Xpred}{A list holding the values of distributional and/or scalar predictors 
for a held-out set awaiting prediction. The form and structure are similar to 
the input \code{X} of \code{\link{WR}}, except that the number of realizations 
awaiting prediction can be different from the number of realizations \eqn{n} used to fit the model.}

\item{...}{Not used.}
}
\value{
A list of the following
\item{Ypred}{Predicted responses; either a vector of length \eqn{n} for scalar responses 
or a \eqn{q}-by-\eqn{n} matrix holding the quantile functions for distributional responses, 
evaluated on \code{qSup}, where \eqn{q} is the length of \code{qSup}.}
\item{qSup}{The input \code{qSup}, only returned when the responses are distributional.}
\item{outOfLogSpace}{A logical vector holding whether initial predicted log mapped distributions 
lie out of the log image space (\code{TRUE}) or not (\code{FALSE}), only returned when the responses are distributional.}
}
\description{
Predict responses for a new sample given a WR object
}
