% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.pwreg}
\alias{print.pwreg}
\title{Print the results of the proportional win-fractions regression model}
\usage{
\method{print}{pwreg}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pwreg}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
print the results of \code{pwreg} object
}
\description{
Print the results of the proportional win-fractions regression model
}
\examples{
# see the example for pwreg
}
\seealso{
\code{\link{pwreg}}
}
