% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.pwreg.score}
\alias{print.pwreg.score}
\title{Print information on the content of the pwreg.score object}
\usage{
\method{print}{pwreg.score}(x, ...)
}
\arguments{
\item{x}{A object of class pwreg.score.}

\item{...}{further arguments passed to or from other methods}
}
\value{
print the results of \code{pwreg.score} object
}
\description{
Print information on the content of the pwreg.score object
}
\examples{
# see the example for score.proc
}
\seealso{
\code{\link{score.proc}}
}
