% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pwreg1.R
\name{print.pwreg1}
\alias{print.pwreg1}
\title{Print model information for a proportional win-fractions regression}
\usage{
\method{print}{pwreg1}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pwreg1}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Print the results of \code{pwreg1} object
}
\description{
Print model information for a proportional win-fractions regression.
}
\examples{
# see the example for pwreg1
}
\seealso{
\code{\link{pwreg1}}
}
\keyword{pwreg1}
