% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.WRrec}
\alias{print.WRrec}
\title{Print the results of the two-sample recurrent-event win ratio analysis}
\usage{
\method{print}{WRrec}(x, ...)
}
\arguments{
\item{x}{an object of class \code{WRrec}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Print the results of \code{WRrec} object.
}
\description{
Print the results of the two-sample recurrent-event win ratio analysis.
}
\examples{
# see the example for WRrec
}
\seealso{
\code{\link{WRrec}}
}
\keyword{WRrec}
