% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pwreg1.R
\name{pwreg1}
\alias{pwreg1}
\title{A test version of pwreg}
\usage{
pwreg1(
  id,
  time,
  status,
  Z,
  wfun = NULL,
  Y = NULL,
  strata = NULL,
  fixedL = TRUE,
  eps = 1e-06,
  maxiter = 50
)
}
\arguments{
\item{time}{A vector of event times.}

\item{status}{A vector of event indicators.}

\item{Z}{A matrix of covariates.}

\item{strata}{A vector of strata indicators.}

\item{fixedL}{A logical value indicating whether the frailty variance is fixed.}

\item{eps}{A numeric value for the convergence criterion.}

\item{maxiter}{A numeric value for the maximum number of iterations.}

\item{ID}{A vector of subject identifiers.}

\item{rho}{A numeric value for the frailty parameter.}
}
\description{
This function is a test version of the \code{pwreg} function.
It is intended to simplify the code and make computations more efficient.
}
