% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFPCAoptns.R
\name{SetFPCAoptns}
\alias{SetFPCAoptns}
\title{Set FPCA option list}
\usage{
SetFPCAoptns(FPCAoptns = list())
}
\arguments{
\item{FPCAoptns}{A list of optional control parameters specified by \code{list(name=value)} for \code{\link[fdapace]{FPCA}}.}
}
\value{
A list of optional control parameters.
}
\description{
Set FPCA option list
}
