% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOptns.R
\name{SetOptns}
\alias{SetOptns}
\title{Set option list for WR and WAR}
\usage{
SetOptns(optns = list(), Ymin = NULL, Ymax = NULL)
}
\arguments{
\item{optns}{A list of optional control parameters specified by \code{list(name=value)} for \code{\link{WR}} or \code{\link{WAR}}.}

\item{Ymin, Ymax}{The minimum and maximum of the input \code{Y} for \code{\link{WAR}} or \code{\link{WR}}, 
the latter if the responses are scalar.}
}
\value{
A list of optional control parameters.
}
\description{
Set option list for WR and WAR
}
