\encoding{UTF-8}
\name{WR}
\alias{WR}
\docType{package}
\title{ Whittaker-Robinson periodogram package }
\description{
  Whittaker-Robinson periodogram for univariate series of quantitative data.
}
\details{
\tabular{ll}{
Package: \tab WR\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2018-10-11\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
  Function \code{WRperiodogram} computes a Whittaker-Robinson periodogram for univariate series of quantitative data.
}

\author{ Pierre Legendre and Guillaume Guénard }

\references{
Legendre, P. & L. Legendre. 2012. Numerical ecology, 3rd English edition. Elsevier Science BV, Amsterdam.
}

\keyword{ package }
