% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score.proc}
\alias{score.proc}
\title{Computes the standarized score processes}
\usage{
score.proc(obj, t = NULL)
}
\arguments{
\item{obj}{an object of class pwreg.}

\item{t}{a vector containing times. If not specified, the function will use
all unique event times from the data.}
}
\value{
An object of class \code{pwreg.score} consisting of \code{t:}
a vector of times; and \code{score:} a matrix whose rows are the standardized score processes
as a function of \code{t}.
}
\description{
Computes the standarized score processes for the covariates.
}
\examples{
library(WR)
head(gbc)
id_unique <-unique(gbc$id)

# Randomly sample 200 subjects from gbc data
set.seed(2021)
id_sample <- sample(id_unique, 200)
gbc_reduce <- gbc[gbc$id \%in\% id_sample, ]

# Use the reduced gbc data for PW analysis
nr <- nrow(gbc_reduce)
p <- ncol(gbc_reduce)-3
ID <- gbc_reduce[,"id"]
time <- gbc_reduce[,"time"]
status <- gbc_reduce[,"status"]
Z <- as.matrix(gbc_reduce[,4:(3+p)],nr,p)
pwreg.obj <- pwreg(time=time,status=status,Z=Z,ID=ID)
print(pwreg.obj)
score.obj <- score.proc(pwreg.obj)
#plot the standardized score process for the first covariate
plot(score.obj, k = 1)
}
\references{
Mao, L. and Wang, T. (2020). "A class of proportional win-fractions
regression models for composite outcomes". Biometrics, 10.1111/biom.13382
}
\seealso{
\code{\link{pwreg}}
}
