% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAR.R
\name{WAR}
\alias{WAR}
\title{Wasserstein autoregressive models.}
\usage{
WAR(
  Q,
  qSup,
  optns = list(),
  n.ahead = 1,
  aic = FALSE,
  order.max = 1,
  method = "yule-walker",
  ...
)
}
\arguments{
\item{Q}{A \eqn{q}-by-\eqn{n} matrix holding the quantile functions of 
\eqn{n} distributions evaluated on a common grid on \eqn{[0,1]} consisting of \eqn{q} points.}

\item{qSup}{A vector of length \eqn{q} holding the support grid of quantile functions.}

\item{optns}{A list of options control parameters specified by list(name=value). See 'Details'.}

\item{n.ahead}{Number of steps ahead at which to predict. Default: 1.}

\item{aic, order.max, method, ...}{arguments for \code{\link[stats]{ar}}. 
Default: \code{aic = FALSE, order.max = 1, method = "yule-walker"}. 
Note that \code{demean} should not be input and is set to be \code{FALSE}.}
}
\value{
A list of class \code{'WAR'} with the following elements:
\item{QPred}{A \eqn{q}-by-\code{n.ahead} matrix holding the quantile functions of the \code{n.ahead} predicted distributions.}
\item{qSup}{The support of quantile functions, same as the input.}
\item{order}{The order of the fitted Wasserstein autoregressive model.}
\item{optns}{The \code{optns} used.}
}
\description{
Wasserstein autoregressive models.
}
\details{
Available control options are
\describe{
\item{methodNFPC}{The method for choosing the number of FPCs; \code{'FVE'} (default), \code{'CV'}.}
\item{methodProj}{The projection method if the predictions of log mapped distributional responses lie out of the log image space: 
\code{'log'} (the method as per Chen, Lin, and Müller, 2021); 
\code{'qt'} (default, the method as per Petersen & Müller, 2019).}
\item{FVEthreshold}{Fraction-of-Variance-Explained threshold used for choosing the number of FPCs, 
relevant only if \code{methodNFPC} is \code{'FVE'}; numeric \eqn{(0,1]}, default: 0.95.}
\item{CVfold}{The number of folds for cross validation, relevant only if \code{methodNFPC} is \code{'CV'}; 
Any positive integer up to \eqn{n}; Default: \eqn{n} (leave-one-out) if \eqn{n\le 30} and \eqn{10} (10-fold) if \eqn{n > 30}.}
\item{lower}{Lower bound of the support of distributions; default: \code{NULL}, i.e., no finite lower bound.}
\item{upper}{Upper bound of the support of distributions; default: \code{NULL}, i.e., no finite upper bound.}
}
}
\references{
\cite{Chen, Y., Lin, Z., & Müller, H.-G. (2021). "Wasserstein regression." Journal of the American Statistical Association, in press.}
\cite{Petersen, A., & Müller, H.-G. (2019). "Fréchet regression for random objects with Euclidean predictors." The Annals of Statistics, 47(2), 691--719.}
}
