% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WR-package.R
\docType{package}
\name{WR-package}
\alias{WR}
\alias{WR-package}
\title{WR: Win Ratio Analysis of Composite Time-to-Event Outcomes}
\description{
Implements various win ratio methodologies for composite endpoints of death and non-fatal events, including the (stratified) proportional win-fractions (PW) regression models (Mao and Wang, 2020 \doi{10.1111/biom.13382}), (stratified) two-sample tests with possibly recurrent nonfatal event, and sample size calculation for standard win ratio test (Mao et al., 2021 \doi{10.1111/biom.13501}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/view/lmaowisc/}
  \item \url{https://lmaowisc.github.io/WR/}
}

}
\keyword{internal}
